fix: address featured market review findings
